refactor: move ProjectService from internal/user to internal/project
- Add internal/project/service_user.go with inline request/decode logic
- Add internal/project/service_user_test.go (ported from internal/user)
- Remove ProjectService and NewProjectService from internal/user/service.go
- Remove internal/user/service_project_test.go
- Update user.go to use project.NewProjectUserService
